Partially created tests for DiseaseHistory domain

+++ src/main/java/com/jupiter/asclepi/core/model/request/disease/history/GetDiseaseHistoryRequest.java
@@ -2,8 +2,10 @@ package com.jupiter.asclepi.core.model.request.disease.history;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class GetDiseaseHistoryRequest {
-    private Integer clientId;
+    private BigInteger clientId;
     private Integer number;
 }
