+++ src/main/java/com/jupiter/asclepi/core/repository/client/ClientRepository.java
@@ -4,6 +4,8 @@ import com.jupiter.asclepi.core.model.entity.people.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, BigInteger> {
 }
