+++ src/main/java/com/jupiter/asclepi/core/rest/controller/impl/ClientControllerImpl.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -67,7 +68,7 @@ public class ClientControllerImpl implements ClientController {
     }
 
     @Override
-    public ResponseEntity<ClientInfo> getOne(@NotNull Integer getRequest) {
+    public ResponseEntity<ClientInfo> getOne(@NotNull BigInteger getRequest) {
         return clientService.getOne(getRequest)
                 .map(client -> conversionService.convert(client, ClientInfo.class))
                 .map(ResponseEntity::ok)
